requirements.txt
flask
pycups
pillow
